colonel_kurtz_ex v0.1.1 ColonelKurtz.BlockType

The BlockType module defines a macro that is used to mix in the Block Type behavior. Block Types are embedded Ecto Schemas.

A Block has a type (e.g. "image"), a list of children (blocks), and a well-defined schema with content attributes (content). The content schema is defined by a user in a TypeName.Content module. The Block Type schema embeds_one of the Content module.

Types

t()

t() :: %atom(){
  block_id: nil | binary(),
  type: binary(),
  content: map(),
  blocks: [t()]
}

Functions

__using__(opts)
(macro)

The BlockType using macro allows modules to behave as BlockTypes.

attributes_from_params(params)

attributes_from_params(map()) :: map()

lift_content_errors(changeset)

lift_content_errors(changeset()) :: changeset()