colonel_kurtz_ex v0.1.0 ColonelKurtz.BlockTypeContent

BlockTypeContent is used to model the inner content field for a given block type. Each block type has a unique schema represented by an Ecto.Schema that defines an embedded_schema for the expected fields (corresponding with the JS implementation for this block type).

Modules that use this macro may optionally define a validate/2 that receives the Content struct and a changeset with the fields in their embedded_schema already cast and ready to be validated.

Types

t()

t() :: %atom(){optional(any()) => any()}

Functions

__using__(opts)
(macro)

Macro for mixing in the BlockTypeContent behavior.

params_map(params)

params_map(map() | struct()) :: map()